Zappos Challenge
This program leverages the Zappos API to find a sequnce of item given
user input on their desired cost and desired number of products they
wish to purchase for the same cost or lower. It is built in Python and
requires no additional libraries. The main idea here is to recursively
compute a list that satisfies the given condition.